create UI based on the data dynamically
Hard code the row data is not scalable. If we have more and more
student, then we have to copy the table row over and over. As a result
I prefer create the table rows using JavaScript.

By the way, this code has the following problems:

1. We have to empty local storage after we add a new student, otherwise
the new student will never be shown.

2. number of days is not changeable.